test and refactor twitter message generation
