+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/TwitterPublishingCommander.scala
@@ -23,6 +23,7 @@ class TwitterPublishingCommander @Inject() (
     libraryImageCommander: LibraryImageCommander,
     userRepo: UserRepo,
     libraryImageStore: LibraryImageStore,
+    twitterMessages: TwitterMessages,
     keepImageStore: KeepImageStore,
     twitterSocialGraph: TwitterSocialGraph) extends SocialPublishingCommander with Logging {
 
@@ -38,7 +39,7 @@ class TwitterPublishingCommander @Inject() (
           }
           val libraryUrl = s"""https://www.kifi.com${Library.formatLibraryPath(libOwner.username, library.slug)}"""
           val title = keep.title.getOrElse("interesting link")
-          val msg = keepMessage(title, keep.url, library.name, libraryUrl)
+          val msg = twitterMessages.keepMessage(title, keep.url, library.name, libraryUrl)
           log.info(s"twitting about user $userId keeping $title with msg = $msg of size ${msg.size}")
           val imageOpt: Option[Future[TemporaryFile]] = keepImageCommander.getBestImageForKeep(keep.id.get, ImageSize(1024, 512)).flatten.map { keepImage =>
             keepImageStore.get(keepImage.imagePath)
@@ -53,28 +54,6 @@ class TwitterPublishingCommander @Inject() (
           }
       }
     }
-  }
-
-  /**
-   * todo(eishay): before openning it out, should be heavily tested.
-   * 140 is twitter max msg length
-   * 20 is the urls after shortning
-   * 20 is the message overhead " kept to via @kifi"...
-   * 140 - 2 * 20 - 20 = 79
-   */
-  private def keepMessage(title: String, keepUrl: String, libName: String, libUrl: String): String = {
-    val contentLength = title.length + libName.length
-    val totalLength = 20 * 2 + 20 + title.length + libName.length
-    if (20 * 2 + 20 + title.length + libName.length <= 140) {
-      s"$title $keepUrl kept to $libName $libUrl via @kifi"
-    } else {
-      val overtext = 79 - (2 * 3) - contentLength //the 3 stands for the "..."
-      val maxLibName = (libName.size - overtext / 3).min(20)
-      val shortLibName = libName.abbreviate(maxLibName)
-      val shortTitle = if (title.size > 79 - shortLibName.size) title.abbreviate(79 - 3 - shortLibName.size) else title
-      s"$shortTitle $keepUrl kept to $shortLibName $libUrl via @kifi"
-    }
-
   }
 
   def publishLibraryMembership(userId: Id[User], library: Library): Unit = {
