+++ kifi-backend/modules/shoebox/test/com/keepit/commanders/UserCommanderTest.scala
@@ -1,29 +1,26 @@
 package com.keepit.commanders
 
+import com.google.inject.Injector
+import com.keepit.abook.{ ABookServiceClient, FakeABookServiceClientImpl, FakeABookServiceClientModule }
 import com.keepit.common.concurrent.FakeExecutionContextModule
 import com.keepit.common.crypto.FakeCryptoModule
-import com.keepit.common.time._
-import com.keepit.curator.FakeCuratorServiceClientModule
-import com.keepit.shoebox.FakeKeepImportsModule
 import com.keepit.common.db.Id
+import com.keepit.common.mail._
+import com.keepit.common.social.FakeSocialGraphModule
+import com.keepit.common.store.FakeShoeboxStoreModule
+import com.keepit.common.time._
+import com.keepit.cortex.FakeCortexServiceClientModule
+import com.keepit.curator.FakeCuratorServiceClientModule
+import com.keepit.model._
+import com.keepit.scraper.{ FakeScrapeSchedulerModule, FakeScraperServiceClientModule }
+import com.keepit.search.FakeSearchServiceClientModule
+import com.keepit.shoebox.FakeKeepImportsModule
+import com.keepit.test.ShoeboxTestInjector
 import org.joda.time.DateTime
 import org.specs2.mutable.Specification
 
-import com.keepit.test.{ ShoeboxTestInjector }
-import com.keepit.model._
-import com.keepit.common.mail._
-import com.keepit.abook.{ FakeABookServiceClientImpl, ABookServiceClient, FakeABookServiceClientModule }
-import com.keepit.common.social.FakeSocialGraphModule
-import com.keepit.search.FakeSearchServiceClientModule
-import com.keepit.scraper.{ FakeScraperServiceClientModule, FakeScrapeSchedulerModule }
-import com.keepit.common.store.FakeShoeboxStoreModule
-
-import play.api.test.Helpers.running
-
-import com.google.inject.Injector
-import com.keepit.cortex.FakeCortexServiceClientModule
-import scala.concurrent.duration.Duration
 import scala.concurrent.Await
+import scala.concurrent.duration.Duration
 
 class UserCommanderTest extends Specification with ShoeboxTestInjector {
